auth_service/requirements.txt
Django>=5.0
djangorestframework>=3.14.0
psycopg2>=2.9.9
djangorestframework-simplejwt>=5.3.1
django-cors-headers>=4.3.1
python-dotenv>=1.0.0 